backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.12.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
python-multipart>=0.0.6
python-socketio>=5.10.0
groq>=0.4.0
opencv-python>=4.8.0
aiortc>=1.5.0
av>=11.0.0
aiofiles>=23.2.0
httpx>=0.25.0
python-dotenv>=1.0.0
redis>=5.0.0
# OCR and PDF processing
pytesseract>=0.3.10
pillow>=10.0.0
PyPDF2>=3.0.0
pdf2image>=1.16.0
